requirements.txt
# OpenBridge Dependencies
# Install with: pip install -r requirements.txt

# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Platform Adapters
python-telegram-bot>=20.6
discord.py>=2.3.0

# Infrastructure
redis>=5.0.0
websockets>=12.0
httpx>=0.25.0

# Terminal/PTY
pexpect>=4.8.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# CLI/UI
click>=8.1.0
rich>=13.7.0
questionary>=2.0.0
jinja2>=3.1.2

# Security
cryptography>=41.0.0
PyJWT>=2.8.0

# Utilities
aiofiles>=23.2.0
python-multipart>=0.0.6
psutil>=5.9.6

# Monitoring
prometheus-client>=0.19.0
structlog>=23.2.0

# Daemon
python-daemon>=3.0.0
